@@ -56,22 +56,22 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "13.1.0",
-    "@angular-devkit/build-angular": "13.3.9",
+    "@angular-devkit/build-angular": "14.2.2",
     "@angular-eslint/builder": "14.0.3",
     "@angular-eslint/eslint-plugin": "14.0.3",
     "@angular-eslint/eslint-plugin-template": "14.0.3",
     "@angular-eslint/schematics": "14.0.3",
     "@angular-eslint/template-parser": "14.0.3",
-    "@angular/animations": "13.3.11",
-    "@angular/cli": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/animations": "14.2.1",
+    "@angular/cli": "14.2.2",
+    "@angular/common": "14.2.1",
+    "@angular/compiler": "14.2.1",
+    "@angular/compiler-cli": "14.2.1",
+    "@angular/core": "14.2.1",
+    "@angular/forms": "14.2.1",
+    "@angular/platform-browser": "14.2.1",
+    "@angular/platform-browser-dynamic": "14.2.1",
+    "@angular/router": "14.2.1",
     "@fortawesome/fontawesome-free": "6.2.0",
     "@playwright/test": "1.25.1",
     "@swimlane/ngx-datatable": "20.0.0",
@@ -117,7 +117,7 @@
     "squel": "5.13.0",
     "ts-mockito": "2.6.1",
     "ts-node": "10.9.1",
-    "typescript": "4.5.5",
+    "typescript": "4.8.3",
     "wait-on": "6.0.1",
     "webdriver-manager": "12.1.8",
     "webpack": "5.74.0",
